Sourcecode_Deploy_Model/requirements.txt
Flask
gunicorn
google-cloud-storage
pillow
numpy
tensorflow
keras
pandas
chardet